separate cxa_init_primary_exception test into its own file
